pyproject.toml
[tool.poetry]
name = "cbotdiscord"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-discord = "^0.1.64"
oauth = "^1.0.1"
pymongo = "^3.11.4"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
